Rewrite enchantment related code. Fixed: Include item in inventory save in all cases when enchantment added/removed or duration/charges changed. Fixed: remove held enchantments at unequip, not double remove enchantment bonuses at equipped item destroy. Fixed: not check item target existence at spell cast (just ignore it at spell casting if not found).This will prevent output error at message at offhand item that apply held enchantment to main hand weapon.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@3673 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/ItemHandler.cpp
@@ -780,10 +780,10 @@ void WorldSession::SendEnchantmentLog(uint64 Target, uint64 Caster,uint32 ItemID
 void WorldSession::SendItemEnchantTimeUpdate(uint64 Playerguid, uint64 Itemguid,uint32 slot,uint32 Duration)
 {
     WorldPacket data(SMSG_ITEM_ENCHANT_TIME_UPDATE, (8+4+4+8)); // last check 2.0.10
-    data << Itemguid;
-    data << slot;
-    data << Duration;
-    data << Playerguid;
+    data << uint64(Itemguid);
+    data << uint32(slot);
+    data << uint32(Duration);
+    data << uint64(Playerguid);
     SendPacket(&data);
 }
